Finish Q268

+++ 1346. 检查整数及其两倍数是否存在/src/main.java
@@ -7,8 +7,8 @@ class Solution1346 {
     public boolean checkIfExist(int[] arr) {
         Set set = new HashSet();
         for (int num : arr) {
-            if (set.contains((float) num*2) || set.contains((float) num/2)) return true;
-            set.add((float) num);
+            if (set.contains(num*2) || set.contains(num/2)) return true;
+            set.add(num);
         }
         return false;
     }
